feat(hyprland): add workspace bindings for most keys on the keyboard
Moving away from using numbers as workspaces as it is easy to mispress
and move to the wrong workspace due to the fact that the number row is
two rows away from the home row. Using other keys feels more ergonomic.
The numbers are still usable as workspaces, but I'll likely rarely use
them.

Signed-off-by: Dixon Sean Low Yan Feng <dixonseanlow@protonmail.com>