revert: "feat(firefox): add extensions (from nur)"
It's easier to just use firefox sync to manage my extensions, settings,
etc. With home-manager's firefox module, extensions don't get enabled
automatically on first run, which isn't too big of a deal. But add in
impermanence, and it becomes quite a problem figuring out which files in
.mozilla to persist. Firefox also doesn't seem to play nice with broken
symlinks. There's also the issue of declaring extension settings.

This reverts commit 9c7566c0b753100a5b6bb14614844af8cfd2ae7f.