refactor(nvim): use vim plugins from nixpkgs if possible
The original reason for maintaining my own package set of vim plugins
was for faster updates and for installing plugins that are not in
nixpkgs quicker.

However, nixpkgs actually updates vim plugins pretty quickly, and has
overrides for more complex plugins like treesitter that are troublesome
to maintain by myself. Consequently, I might as well use the ones from
nixpkgs. It would be better to contribute packages that are missing too,
instead of keeping them for myself.

Signed-off-by: Dixon Sean Low Yan Feng <dixonseanlow@protonmail.com>