feat(services): add syncthing configuration
Configuration files are encrypted using sops, then decrypted during
activation. Symlinks are made to the decrypted files. Because they are
outside of the nix store, it's not as straightforward to create symlinks
to them. sops-nix can create symlinks, but as described in issue 235,
the directory permissions are set to root, which breaks the whole
system.

The NixOS module is nice in that it provides a way to specify the
syncthing configuration declaratively, but doing that would leak
sensitive information such as device IDs. Unfortunately, with sops-nix,
secrets cannot be used at evaluation time, so that's a no-go. The module
doesn't provide a way to make the syncthing service a user service as
well.

Signed-off-by: Dixon Sean Low Yan Feng <dixonseanlow@protonmail.com>