fix(nvim, lsp): pass separate capabilities to each server
Lua passes tables by *reference*, so modifying the capabilities in any
of the server configurations would actually modify the capabilities for
all of the servers. This, in addition to setting the `offsetEncoding`
incorrectly (should be an array, not a string; will have to fix this as
well) for clangd, resulted in rust-analyzer not working. Formatting
probably got disabled for all of the servers as well since it's disabled
for some of the servers.