refactor(pkgs): reduce verbosity via `mapAttrs`
Uses `mapAttrs` to map sources to `callPackage`.
Packages that don't use npins are defined in a separate set.
Rather than using `callPackage` to be able to reference the other
packages, the dependencies are explicitly passed to the `callPackage`.
I think this is clearer.

`catppuccin-rofi-basic` has been renamed to `catppuccin-rofi` to
match the source name. I would prefer to keep the original name, but
honestly it's not a big deal.