refactor(polybar): use systemd services for monitor-specific polybars
A systemd service is created for each monitor's polybar, which are both
`WantedBy` `bspwm-session.target`. `autorandr` has not yet been
configured to use these services in this commit.

Services are also used to set the top margin for bspwm, which allows me
to specify a dependency between the polybar services and the margins.
The margins shouldn't be set in the polybar services because polybar may
not always run on bspwm (though that *is* the case now since I don't have
any other X WMs configured); the margins are specific to bspwm.