ci: add action to build and push to cachix
Adds an action that builds the specified flake output and pushes the
result to cachix. Optionally, an artifact of the build outputs will be
uploaded, which is useful for the deploy spec.

ci: don't checkout, but allow specifying flake path for the `build` action

ci: specify `shell` to fix error in the `build` action

ci: set artifact path correctly in the `build` action

Apparently, `working-directory` doesn't work with `uses`.
Yet another instance of lack of documentation.

ci: don't use `flake-path` through `working-directory`

ci: add option to skip cachix push in the `build` action

Pushing has to be manually done in `deploy`, otherwise the push is only
done after the deployment. In that scenario, there's no point pushing
from this action.

ci: don't upload artifact in the `build` action

Seems out of scope for this action.