ci(cachix): ignore workflow conclusion when downloading artifact
The workflow doesn't necessarily need to be successful for the artifact
to be available. For example, if `build-others` fails but the rest of
the jobs succeed, then the artifact will still be available.

More importantly, `action-download-artifact` only falls back to
`download-artifact` action if the workflow conclusion is empty (ignored)
or `in_progress`. By default, only `success`ful workflows are searched
(poor default IMO), which meant that artifacts from the same run could
not be downloaded, a behaviour that is necessary to build and deploy in
the same run.