ci(update): use single job for npins
Uses `jq` to generate a list of objects that will act as the matrix.

Each object has a `name`, which is the name of the pin, and an
`npins-dir`, which is the npins base directory that contains the pin
entry.

With this matrix, a separate job for vim plugins is no longer needed.

ci(update-npins): rename `npins-path` to `npins-dir`

More consistent with what npins uses.

ci(update): commit and push if pin is updated

ci(update): add `-c` flag to jq to print compactly

ci(update): use `install-nix-action` `v19` for `update-flake`