refactor(flake): introduce `home-manager` flake module
This flake module will contain my home-manager configuration, and is a
first step in a major reorganization to make it easier to share
configuration between multiple hosts (something I'm thinking of doing
in the near future).

A bonus is that I can now build the home-manager configuration separately
via `homeConfigurations.<user>.activationPackage` (broken for now), which
may be useful in CI for build parallelization (need to search more about
this though).