fix(hm): apply overlays to `nixpkgs`
Applies the required overlays to `nixpkgs`.

Since the NixOS configuration already applies overlays, the overlays are
only applied to the HM configuration so that the overlays aren't
duplicated. The NixOS configuration merges the overlays that it needs
for the system configuration with the overlays that the HM configuration
requires. Then, `lib.lists.unique` is applied to ensure that overlays
required for both system and HM configuration are not duplicated.