feat(planet, hm, persist): introduce `persistence` module
Adds a new `persistence` module to `planet`. It is essentially a
wrapper around the the `impermanence` HM module, but slightly more
convenient in that the persistent storage location is already set by
default, so there's no need to keep typing it out, which helps in
reducing the changes of a typo. The downside though is that only one
persistent storage location can be set; it's probably not too difficult
to add this in the future if I need to, however, and it's not likely
that I'll need to be able to do that.

There's also an option to easily persist the XDG user directories for
convenience.