fix(sops): use age key for sops home-manager decryption
Previously, the `sops-nix` user service was failing as it was unable to
decrypt some files on login. I think this is because gpg requires me to
enter my password, but that doesn't happen on login.

So, a new age key has been generated and added to `.sops.yaml` and those
files to fix this.