revert: "refactor(planet, hm, nixos): don't assume that `inputs` is passed by the hm / nixos configuration"
This reverts commit 9edc8da2dabcd0cd4ec1d6cd7d9ad2ab3bb271df.

This won't work because using values from `_modules.args` in `imports`
will cause infinite recursion. `imports` needs to be evaluated to reach
the `config` fixpoint, but to evaluate `inputs`, we need the `config`
fixpoint (because `_module.args` is an option assignment in `config`),
so we need the `config` fixpoint to determine the `config` fixpoint
i.e., a paradox.