refactor(flake, planet, hm, neovim): replace `oxalica/rust-overlay` with `fenix`
`oxalica`'s overlay is nice, but there's no easy way to access the rust
toolchain as a package. It also seems to include old versions of the
toolchain, so it's quite a bit large. `fenix` provides both an overlay
and packages, and does not include old versions, so it's smaller.

Not requiring an overlay is important for `planet` (`nvim` in
particular) because 1) I don't want to silently use the `overlay` in
`planet`, and 2) I don't want an overlay to be a requirement of using
`planet`. In fact, overlays in general can be hard to keep track of.
It's much easier to reason if a package from an external flake is used
directly.