ci(cachix): don't use `build` composite action
The `build` composite action was actually complicating things a bit.
For example, `build-deploy-specs` needs to install `nix` to import the
store cache, but the `build` composite action also installs `nix`, so
the `install-nix-action` runs twice.

The composite action was originally created because the `deploy` job
used to use it. However, that is no longer the case, so it's actually
obsolete now.