ci(build): delete as it is obsolete
As mentioned in the previous commit, the `build` composite action is now
obsolete. It was created because the build and deploy jobs used to have
a lot of shared steps since deploying could actually build as well.
However, the deploy job now does not do any building now, so the
composite action is no longer needed.

It'll still be a good idea to see what steps can be factored out into
another composite action though.