ci(cachix): add ability to force rebuild
In some cases, the deployment specification artifact exists, but for
some reason, the build outputs are not cached on cachix (probably
something to do with garbage collection on cachix's end, but I'm not
sure). This commit makes it possible to skip the check for the
deployment specification artifact to force a rebuild and cache the build
outputs.

[skip ci]