fix(planet/hm/neovim): #18
Fixes #18.

The `vim` and `vi` aliases were enabled using the NixOS `neovim` module.
I didn't dig too deep, but it seems that the home-manager module uses a
wrapper to "install" the plugins. The aliases were hence calling the
NixOS neovim, and not the home-manager neovim. This commit fixes that by
enabling the aliases in the home-manager neovim module (it also enables
the `defaultEditor`, which is unrelated and enabled for good measure).
The corresponding options in the NixOS module are left enabled as the
`root` user can take advantage of those.