fix(planet/hm/sxhkd): execute with a login shell
Previously, many environment variables were missing in `sxhkd.service`,
which led to many issues. For example, `fish` complained about missing
binaries. I imagine that other programs were affected by the missing
environment variables as well.

This commit fixes that by starting `sxhkd` with a login shell. Details
about bash's behaviour when started as a login shell can be found here:
https://www.gnu.org/software/bash/manual/html_node/Bash-Startup-Files.html.
Importantly, when a shell is started as a login shell, it executes
certain files (`/etc/profile` and others), which are where most, if not
all, required environment variables (e.g., `PATH`) are set. Starting
`sxhkd` with a login shell (`/bin/sh`, which exists on NixOS for POSIX
compatibility if I'm not wrong) populates the environment variables.

This solution feels hacky, but I can't think of a cleaner and more
reliable way to set the environment variables correctly without having
to manually specify them in the `sxhkd` service file. Manually
specifying them isn't feasible as they are bound to change as the
dotfiles are modified.