ci(cachix): reduce max jobs to 2
It seems like the GitHub Actions runner sometimes exits for whatever
reason. I suspect it's caused by either using too much CPU or memory.

By default, the max jobs is set to auto (by Cachix's
install-nix-action), which is equivalent to the number of cores in the
system (Linux GitHub Actions runners have 4 cores). This commit forces
it to 2 to hopefully resolve the issue.