refactor(nixos/delta): mount esp on `/boot`
Seems like systemd-boot can only start EFI executables on the esp or an
XBOOTLDR partition. The filesystem the XBOOTLDR partition uses must be
supported by the UEFI implementation; the supported ones are usually
FAT (since that is the only one required by the UEFI spec), so there's
not much point in having a separate XBOOTLDR partition in my case
(since the ESP itself needs to be FAT). This does mean that I can't use
an encrypted boot partition with systemd-boot, but that's not that big
of an issue (I could look at secure boot if that level of security is
required).